requirements/was_requirements.txt
bleach==2.1.2
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
Django==1.11
django-extensions==1.9.9
django-ipware==2.0.1
django-markdown-deux==1.0.5
django-rest-framework==0.1.0
django-rest-multiple-models==2.0.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
entrypoints==0.2.3
html5lib==1.0.1
idna==2.6
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
Markdown==2.6.11
markdown2==2.3.5
MarkupSafe==1.0
mistune==0.8.3
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.4
requests==2.18.4
rest-framework-generic-relations==1.1.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
